added logs header, added Cube drawing code

+++ src/actors/Cube.h
@@ -5,8 +5,8 @@
 
 class Cube : public Actor {
 private:
-    static const float VERTICES[];
-    static const float INDEXES[];
+    static const GLfloat VERTICES[];
+    static const GLuint INDEXES[];
     static const float TEXTURE_COORDS[];
 
 public:
